requirements.txt
streamlit
python-pptx
google-genai
python-dotenv
